Change FPTree to be backed by a Vec<Vec<FPNode>>.
This means nodes can track their parents, children, and the i-lists using
id's (indicies) and so the borrow checker is easier to handle. We don't
need to walk the tree while constructing conditional trees in order to
create a lookup table for the parents or the i-list anymore; we can
maintain it while building the FPTree.